refactor(mentor): constants 모듈화 및 모듈 정리
